requirements.txt
# Semantic Diffusion Model Training Requirements
# Compatible with A100 GPUs and distributed training

# Core dependencies
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
numpy==1.24.3
pillow==10.0.0

# Hugging Face ecosystem
transformers==4.33.2
datasets==2.14.4
accelerate==0.22.0
diffusers==0.20.2

# Image processing
opencv-python==4.8.0.76
scikit-image==0.21.0
imageio==2.31.3
matplotlib==3.7.2
seaborn==0.12.2

# Distributed training
mpi4py==3.1.4
deepspeed==0.10.3

# Data handling
pandas==2.0.3
h5py==3.9.0
tqdm==4.66.1

# Evaluation metrics
lpips==0.1.4
pytorch-fid==0.3.0

# Logging and monitoring
tensorboard==2.14.0
wandb==0.15.8

# Utilities
click==8.1.7
omegaconf==2.3.0
hydra-core==1.3.2